Django/Beargrass_Christian_Church/requirements.txt
django-debug-toolbar
markdown2
channels==1.1.8
Django==1.11.7
djangorestframework==3.7.3
